--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-field-group-wrapper/dynamic-field-group-wrapper.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { FieldWrapper } from '@ngx-formly/core';
-
-@Component({
-  selector: 'op-dynamic-field-group-wrapper',
-  templateUrl: './dynamic-field-group-wrapper.component.html',
-  styleUrls: ['./dynamic-field-group-wrapper.component.scss'],
-  standalone: false,
-})
-export class DynamicFieldGroupWrapperComponent extends FieldWrapper {
-}
