--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/boolean-input/boolean-input.component.html
@@ -1,6 +1,0 @@
-<input
- type="checkbox"
- [attr.aria-required]="to.required"
- [formControl]="formControl"
- [formlyAttributes]="field"
->
