--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/boolean-input/boolean-input.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { FieldType } from '@ngx-formly/core';
-
-@Component({
-  selector: 'op-boolean-input',
-  templateUrl: './boolean-input.component.html',
-  styleUrls: ['./boolean-input.component.scss'],
-  standalone: false,
-})
-export class BooleanInputComponent extends FieldType {
-}
