--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/date-input/date-input.component.html
@@ -1,5 +1,0 @@
-<op-basic-single-date-picker [required]="to.required"
-  [disabled]="to.disabled"
-  [formControl]="formControl"
-  [formlyAttributes]="field"
- />
