--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/date-input/date-input.component.ts
@@ -1,17 +1,0 @@
-import { ChangeDetectionStrategy, Component, HostBinding } from '@angular/core';
-import { FieldType } from '@ngx-formly/core';
-
-@Component({
-  selector: 'op-date-input',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  templateUrl: './date-input.component.html',
-  styleUrls: ['./date-input.component.scss'],
-  standalone: false,
-})
-export class DateInputComponent extends FieldType {
-  @HostBinding('class') get class() {
-    return (this.model?.id === 'projects' && this.key.toString().startsWith('customField'))
-      ? 'form--date-picker-container -xslim'
-      : null;
-  }
-}
