--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/formattable-textarea-input/components/formattable-control/formattable-control.component.html
@@ -1,8 +1,0 @@
-<div class="op-ckeditor--wrapper op-ckeditor-element">
-  <op-ckeditor [context]="ckEditorContext"
-      [content]="value?.raw"
-      (contentChanged)="onContentChange($event)"
-      (initializationFailed)="initializationError = true"
-      (initializeDone)="onCkeditorSetup($event)"
-   />
-</div>
