--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/formattable-textarea-input/formattable-textarea-input.component.html
@@ -1,4 +1,0 @@
-<op-formattable-control [templateOptions]="to"
-  [formControl]="formControl"
-  [formlyAttributes]="field"
- />
