--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/formattable-textarea-input/formattable-textarea-input.component.scss
@@ -1,3 +1,0 @@
-:host {
-  overflow: hidden;
-}
