--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/formattable-textarea-input/formattable-textarea-input.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { FieldType } from '@ngx-formly/core';
-
-@Component({
-  selector: 'op-formattable-textarea-input',
-  templateUrl: './formattable-textarea-input.component.html',
-  styleUrls: ['./formattable-textarea-input.component.scss'],
-  standalone: false,
-})
-export class FormattableTextareaInputComponent extends FieldType {
-}
