--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/integer-input/integer-input.component.html
@@ -1,8 +1,0 @@
-<input
-   [type]="to.type"
-   [attr.aria-required]="to.required"
-   [attr.lang]="to.locale"
-   [formControl]="formControl"
-   [formlyAttributes]="field"
-   class="op-input"
->
