--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/integer-input/integer-input.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { FieldType } from '@ngx-formly/core';
-
-@Component({
-  selector: 'op-integer-input',
-  templateUrl: './integer-input.component.html',
-  styleUrls: ['./integer-input.component.scss'],
-  standalone: false,
-})
-export class IntegerInputComponent extends FieldType {
-}
