--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/project-input/project-input.component.html
@@ -1,6 +1,0 @@
-<op-project-autocompleter [formControl]="formControl"
-  [formlyAttributes]="field"
-  [attr.aria-required]="to.required"
-  [attr.required]="to.required"
-  [url]="to.allowedValuesHref"
- />
