--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/project-input/project-input.component.ts
@@ -1,14 +1,0 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-} from '@angular/core';
-import { FieldType } from '@ngx-formly/core';
-
-@Component({
-  selector: 'op-project-input',
-  templateUrl: './project-input.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  standalone: false,
-})
-export class ProjectInputComponent extends FieldType {
-}
