--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/text-input/text-input.component.html
@@ -1,7 +1,0 @@
-<input
-  [type]="to.type"
-  [attr.aria-required]="to.required"
-  [formControl]="formControl"
-  [formlyAttributes]="field"
-  class="op-input"
->
