--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/text-input/text-input.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { FieldType } from '@ngx-formly/core';
-
-@Component({
-  selector: 'op-text-input',
-  templateUrl: './text-input.component.html',
-  styleUrls: ['./text-input.component.scss'],
-  standalone: false,
-})
-export class TextInputComponent extends FieldType {
-}
