--- frontend/src/app/shared/components/dynamic-forms/components/dynamic-inputs/user-input/user-input.component.html
@@ -1,8 +1,0 @@
-<op-user-autocompleter [formControl]="formControl"
-  [formlyAttributes]="field"
-  [attr.aria-required]="to.required"
-  [attr.required]="to.required"
-  [url]="to.allowedValuesHref"
-  [multiple]="to.multiple"
-  [inviteUserToProject]="projectId"
- />
